refactor: Fix more imports in moved tests
